+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Inspired from [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ## [Unreleased 2.x]
 
 ### Added
+- Add aggregations and suggest in SearchTemplateResponse ([#456](https://github.com/opensearch-project/opensearch-java/issues/456))
 
 ### Dependencies